% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/proto.R
\docType{package}
\name{proto-package}
\alias{proto-package}
\title{Object-Oriented Programming with the Prototype Model}
\description{
Object-oriented programming with the prototype model.  \code{"proto"}
facilitates object-oriented programming using an approach that emphasizes
objects rather than classes (although it is powerful enough to readily
represent classes too).
}
\examples{
cat("parent\\n")
oop <- proto(x = 10, view = function(.) paste("this is a:", .$x))
oop$ls()
oop$view()

cat("override view in parent\\n")
ooc1 <- oop$proto(view = function(.) paste("this is a: ***", .$x, "***"))
ooc1$view()

cat("override x in parent\\n")
ooc2 <- oop$proto(x = 20)
ooc2$view()
}
\keyword{programming}

